Move Users resource to v1
