+++ config/routes.rb
@@ -1,5 +1,8 @@
 Rails.application.routes.draw do
   namespace :api do
-    resources :users, :transactions
+    resources :transactions
+  end
+  namespace :v1 do
+    resources :users
   end
 end
